podcasts/yandex/requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
docutils==0.21.2
frozenlist==1.4.1
idna==3.7
lockfile==0.12.2
magic-filter==1.0.12
multidict==6.0.5
mutagen==1.47.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PySocks==1.7.1
python-daemon==3.0.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==71.0.4
typing_extensions==4.12.2
urllib3==2.2.2
yandex-music==2.2.0
yarl==1.9.4
